% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plot_centrality}
\alias{plot_centrality}
\title{Plot centrality measures and 95\% highest density interval}
\usage{
plot_centrality(output, measure = "Strength")
}
\arguments{
\item{output}{Output object from the bgm_extract function}

\item{measure}{Centrality measures that should be plotted. Users can choose "all" or a subsection of the list: "Strength", "Closeness", "Betweenness", or "ExpectedInfluence"}
}
\description{
Plot centrality measures and 95\% highest density interval
}
