% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plot_edgeevidence}
\alias{plot_edgeevidence}
\title{Edge evidence plot}
\usage{
plot_edgeevidence(output, evidence_thresh = 10, split = F, show = "all", ...)
}
\arguments{
\item{output}{Output object from the bgm_extract function}

\item{evidence_thresh}{BF which will be considered sufficient evidence for in-/exclusion}

\item{split}{if TRUE, plot is split in included and excluded edges}

\item{show}{specifies which edges should be shown, indicated by "all", "included", "inconclusive", "excluded"}

\item{...}{Additional \code{qgraph} arguments}
}
\description{
Edge evidence plot
}
