% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plot_network}
\alias{plot_network}
\title{Network plot}
\usage{
plot_network(output, exc_prob = 0.5, dashed = F, ...)
}
\arguments{
\item{output}{Output object from the bgm_extract function}

\item{exc_prob}{threshold for excluding edges; all edges with a lower inclusion probability will not be shown}

\item{dashed}{binary parameter indicating whether edges with inconclusive evidence should be dashed}

\item{...}{Additional \code{qgraph} arguments}
}
\description{
Network plot
}
