% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plot_parameterforest}
\alias{plot_parameterforest}
\title{Parameter forest plot of interaction parameters and 95\% highest density intervals}
\usage{
plot_parameterforest(output)
}
\arguments{
\item{output}{Output object from the extract_results function}
}
\description{
Parameter forest plot of interaction parameters and 95\% highest density intervals
}
