% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plot_parameterHDI}
\alias{plot_parameterHDI}
\title{Plot of interaction parameters and their 95\% highest density intervals}
\usage{
plot_parameterHDI(output)
}
\arguments{
\item{output}{Output object from the bgm_extract function}
}
\description{
Plot of interaction parameters and their 95\% highest density intervals
}
