% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plot_posteriorcomplexity}
\alias{plot_posteriorcomplexity}
\title{Plot posterior structure complexity}
\usage{
plot_posteriorcomplexity(output)
}
\arguments{
\item{output}{Output object from the bgm_extract function}
}
\description{
Plot posterior structure complexity
}
