% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plot_posteriorstructure}
\alias{plot_posteriorstructure}
\title{Plot Posterior Structure Probabilities}
\usage{
plot_posteriorstructure(output, as.BF = FALSE)
}
\arguments{
\item{output}{Output object from the bgm_extract function}

\item{as.BF}{if TRUE plots the y-axis as Bayes factor instead of posterior structure probability}
}
\description{
Plot Posterior Structure Probabilities
}
