% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plot_structure}
\alias{plot_structure}
\title{Structure plot}
\usage{
plot_structure(output, ...)
}
\arguments{
\item{output}{Output object from the bgm_extract function}

\item{...}{Additional \code{qgraph} arguments}
}
\description{
Structure plot
}
