% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgm_extract.R, R/bgm_extract_class.R
\name{bgm_extract}
\alias{bgm_extract}
\title{Extract results from a Bayesian analysis of networks}
\usage{
bgm_extract(fit, ...)

bgm_extract(fit, ...)
}
\arguments{
\item{fit}{Fit object with a particular class that will dispatch to the respective package functions}

\item{...}{Additional arguments to be passed onto the respective fitting functions}

\item{model}{type of model estimated, e.g., ggm, gcgm, dgm-binary, Ising, ordinal}

\item{package}{package used to obtain the fit object}

\item{posterior_samples}{if TRUE, the posterior samples will be extracted. Note will significantly increase the computation time for 'BDgraph'.}

\item{not.cont}{only if model = "gcgm" vector indicating the not-continuous variables}

\item{data}{if posterior_samples = T, provide the raw data used to estimate the network}

\item{centrality}{if TRUE, the centrality samples will be extracted. Note will significantly increase the computation time.}
}
\description{
Extract results from a Bayesian analysis of networks

Extract the results of a Bayesian analysis of networks
}
