% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybgm.R
\name{easybgm}
\alias{easybgm}
\title{Fit a Bayesian analysis of networks}
\usage{
easybgm(
  data,
  type,
  package = NULL,
  not.cont = NULL,
  iter = 10000,
  save = FALSE,
  centrality = FALSE,
  progress = TRUE,
  edge.prior = 0.5,
  ...
)
}
\arguments{
\item{data}{An n x p matrix or dataframe containing the variables for n independent observations on p variables.}

\item{type}{What is the data type? Options: continuos, mixed, ordinal, binary}

\item{package}{The R-package that should be used for fitting the network model. Optional argument;
default values are specified depending on the datatype.}

\item{not.cont}{If data-type is mixed, a vector of length p, specifying the not-continuous
variables (1 = not continuous, 0 = continuous)}

\item{iter}{number of iterations for the sampler}

\item{save}{Logical. Should the posterior samples be obtained (default = FALSE)?}

\item{centrality}{Logical. Should the centrality measures be extracted (default = FALSE)? Note, that it will significantly increase the computation time.}

\item{progress}{Logical. Should a progress bar be shown (default = TRUE)?}

\item{edge.prior}{Default is 0.5. Single value or p x p matrix with one value per edge.}

\item{...}{Additional arguments that are handed to the fitting functions of the packages, e.g., informed prior specifications.}
}
\value{
The returned object of \code{easybgm} contains a lot of information that
is used for visualizing and interpreting the results of a Bayesian analysis of network.
The output includes:

\itemize{

\item \code{parameters} A p x p matrix containing partial associations.

\item \code{inc_probs} A p x p matrix containing the posterior inclusion probabilities.

\item \code{BF} A p x p matrix containing the posterior inclusion Bayes factors.

\item \code{structure} Adjacency matrix of the median probability model (i.e., edges with a posterior probability larger 0.5).
}

In addition, for \code{BDgraph} and \code{bgms}, the function returns:

\itemize{

\item \code{structure_probabilities} A vector containing the posterior probabilities of all visited structures, between 0 and 1.

\item \code{graph_weights} A vector containing the number of times a particular structure was visited.

\item \code{sample_graphs} A vector containing the indexes of a particular structure.
}

For all packages, when setting \code{save = TRUE} and \code{centrality = TRUE}, the function will return the following objects respectively:

\itemize{

\item \code{samples_posterior} A k x iter matrix containing the posterior samples for each parameter (i.e., k = p/(p-1)) at each iteration (i.e., iter) of the sampler.

\item \code{centrality} A p x iter matrix containing the centrality of a node at each iteration of the sampler.
}
}
\description{
Easy estimation of a Bayesian analysis of networks to obtain conditional (in)dependence relations between several variables.
}
\examples{
\donttest{

library(easybgm)
library(bgms)

data <- Wenchuan

# Fitting the Wenchuan PTSD data

fit <- easybgm(data, type = "ordinal",
                iter = 1000 # note for demonstrative purposes, normally 1e5 is recommended
                )

# To extract the posterior parameter distribution
# and centrality measures

fit <- easybgm(data, type = "ordinal",
                iter = 1000, save = TRUE,
                centrality = TRUE)
}
}
